windows stuff
Former-commit-id: f1e8ac39e33bce82dd833d6c2cb5ef37602b93b6 [formerly 2b04db1e37c2c4486442de0ee96aa87d984d0c5c]
Former-commit-id: 0a80e5fbd09cd94789ad29ff8b3083e9551afb8e

+++ Scene/MeshRenderer.cpp
@@ -28,10 +28,10 @@ void MeshRenderer::Material(shared_ptr<::Material> m) {
 bool MeshRenderer::Batchable(Device* device) {
 	auto shader = mMaterial->GetShader(device);
 	if (mNeedsLightData == 2) {
-		mNeedsLightData = shader->mDescriptorBindings.count("Lights");
+		mNeedsLightData = (uint32_t)shader->mDescriptorBindings.count("Lights");
 		if (mNeedsLightData == 1) mLightCountRange = shader->mPushConstants.at("LightCount");
 	}
-	if (mNeedsObjectData == 2) mNeedsObjectData = shader->mDescriptorBindings.count("Objects");
+	if (mNeedsObjectData == 2) mNeedsObjectData = (uint32_t)shader->mDescriptorBindings.count("Objects");
 	return mNeedsLightData == 1 && mNeedsObjectData == 1;
 }
